requirements.txt
Django
django-ninja
libvirt-python